♻️ core: move options to its own folder
