+++ packages/classic/src/component/symbols.ts
@@ -1,5 +1,2 @@
 export const $componentMap = Symbol('componentMap');
 export const $componentCount = Symbol('componentCount');
-export const $createStore = Symbol('createStore');
-export const $withContext = Symbol('withContext');
-export const $withStore = Symbol('withStore');
