--- packages/classic/src/hooks/symbols.ts
@@ -1,3 +1,0 @@
-export const $onAdd = Symbol('onAdd');
-export const $onReset = Symbol('onReset');
-export const $onSet = Symbol('onSet');
