+++ packages/classic/src/options/options.ts
@@ -1,8 +1,9 @@
-import { $onReset, $onSet } from '../hooks/symbols';
+import { Component } from '../component/types';
+import { PrefabNode } from '../prefab/types';
+import { $createStore, $onReset, $onSet } from '../options/symbols';
 import { defineHiddenProperty } from '../utils/defineHiddenProperty';
 import { World } from '../world/types';
-import { $createStore } from './symbols';
-import { Component, OnAddFn, WithContextFn, WithStoreFn } from './types';
+import { WithContextFn, WithStoreFn } from './types';
 
 export function withContext<Context>(context: Context): WithContextFn<Context> {
 	return ((component: Component) => {
@@ -26,4 +27,16 @@ export function withStore<Store, Params = unknown, W extends World = World>(
 		if (options.onSet) defineHiddenProperty(component, $onSet, options.onSet);
 		if (options.onReset) defineHiddenProperty(component, $onReset, options.onReset);
 	}) as WithStoreFn<Store, Params>;
+}
+
+export function withParams<P extends PrefabNode>(
+	prefab: P,
+	params: P extends PrefabNode<infer Params> ? Params : never
+): [P, P extends PrefabNode<infer Params> ? Params : never];
+export function withParams<C extends Component>(
+	component: C,
+	params: C extends Component<any, infer Params> ? Params : never
+): [C, C extends Component<any, infer Params> ? Params : never];
+export function withParams(a: any, b: any) {
+	return [a, b];
 }
