+++ packages/classic/src/options/symbols.ts
@@ -1,0 +1,6 @@
+export const $onAdd = Symbol('onAdd');
+export const $onReset = Symbol('onReset');
+export const $onSet = Symbol('onSet');
+export const $createStore = Symbol('createStore');
+export const $withContext = Symbol('withContext');
+export const $withStore = Symbol('withStore');
